Changed the output from the search, and reorganized the layout of the application.
